Refactored Dollar fetching
